requirements.txt
# Core Gmail API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# HTML parsing and processing
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration and environment management
python-dotenv>=1.0.0
pydantic>=2.0.0

# Scheduling
APScheduler>=3.10.0

# Enhanced UI and logging
rich>=13.0.0

# Date/time handling
python-dateutil>=2.8.0

# Performance and reliability (optional but recommended)
requests>=2.28.0
urllib3>=1.26.0

# Development and testing (optional)
# pytest>=7.0.0
# pytest-cov>=4.0.0